pyproject.toml
[project]
name = "agents-repo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-adk>=1.18.0",
    "python-multipart>=0.0.20",
]
